Rename pipe-related files
