+++ src/core/pipe/pipe.js
@@ -1,0 +1,12 @@
+import filter from './pipe.filter';
+import pagination from './pipe.pagination';
+
+export default class Pipe {
+	static get filter() {
+		return filter;
+	}
+
+	static get pagination() {
+		return pagination;
+	}
+}
