+++ src/core/pipe/pipe.pagination.js
@@ -1,0 +1,8 @@
+export default function (data, ctx, next) {
+	let model = ctx.model,
+		pagination = model.pagination(),
+		size = pagination.size,
+		start = pagination.current * size;
+
+	next(data.slice(start, start + size));
+}
